move sections to constants folder
